+++ HomeServiceTracker/Server/Services/ServiceProviderInfo/ServiceProviderInfoService.cs
@@ -12,13 +12,17 @@ namespace HomeServiceTracker.Server.Services.ServiceProviderInfo
             _context = context;
         }
 
+        private Guid _userId;
+        public void SetUserId(Guid userId) => _userId = userId;
+
         public async Task<bool> CreateServiceProviderInfoAsync(ServiceProviderInfoCreate model)
         {
             if (model == null)
                 return false;
             var serviceProviderInfoEntity = new HomeServiceTracker.Server.Models.ServiceProviderInfo
             {
-                ServiceProviderName = model.ServiceProviderName
+                ServiceProviderName = model.ServiceProviderName,
+                OwnerId = _userId
             };
             _context.ServiceProviders.Add(serviceProviderInfoEntity);
             var numberOfChanges = await _context.SaveChangesAsync();
@@ -27,7 +31,8 @@ namespace HomeServiceTracker.Server.Services.ServiceProviderInfo
 
         public async Task<IEnumerable<ServiceProviderInfoListItem>> GetAllServiceProviderInfosAsync()
         {
-            var serviceProviderInfoQuery = _context.ServiceProviders.Select(entity => new ServiceProviderInfoListItem
+            var serviceProviderInfoQuery = _context.ServiceProviders.Where(o => o.OwnerId == _userId)
+                .Select(entity => new ServiceProviderInfoListItem
             {
                 Id = entity.Id,
                 ServiceProviderName = entity.ServiceProviderName,
@@ -38,7 +43,7 @@ namespace HomeServiceTracker.Server.Services.ServiceProviderInfo
 
         public async Task<ServiceProviderInfoDetail> GetServiceProviderInfoByIdAsync(int serviceProviderInfoId)
         {
-            var serviceProviderInfoEntity = await _context.ServiceProviders.FirstOrDefaultAsync(s => s.Id == serviceProviderInfoId);
+            var serviceProviderInfoEntity = await _context.ServiceProviders.FirstOrDefaultAsync(s => s.Id == serviceProviderInfoId && s.OwnerId == _userId);
 
             if (serviceProviderInfoEntity is null)
                 return null;
@@ -59,7 +64,7 @@ namespace HomeServiceTracker.Server.Services.ServiceProviderInfo
             var entity = await _context.ServiceProviders.FindAsync(model.Id);
 
             // NEED TO ASSIGN A PERMISSION TO EDIT A SERVICE PROVIDER
-            //if (entity?.PrimaryHomeownerId != _userId) return false;
+            if (entity?.OwnerId != _userId) return false;
 
             entity.ServiceProviderName= model.ServiceProviderName;
 
@@ -71,7 +76,7 @@ namespace HomeServiceTracker.Server.Services.ServiceProviderInfo
             var entity = await _context.ServiceProviders.FindAsync(serviceProviderInfoId);
 
             // NEED TO ASSIGN A PERMISSION TO EDIT A SERVICE PROVIDER
-            //if (entity?.PrimaryHomeownerId != _userId) return false;
+            if (entity?.OwnerId != _userId) return false;
 
             _context.ServiceProviders.Remove(entity);
             return await _context.SaveChangesAsync() == 1;
